requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
cattrs==24.1.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.4.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.15.0
pandas==2.2.2
platformdirs==4.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
